abstracting to resolve  circular dependency
